Factor out casting.rs
